+++ Implementation Phase/GELF/src/main/java/gelf/view/diagrams/components/HistogramBar.java
@@ -3,8 +3,8 @@ package gelf.view.diagrams.components;
 import java.awt.Color;
 
 public class HistogramBar extends DiagramBar {
-	protected HistogramBar(Color color, double value, PositionIn2DDiagram topLeft, PositionIn2DDiagram bottomRight,
-			double borderThickness) {
+	protected HistogramBar(Color color, float value, PositionIn2DDiagram topLeft, PositionIn2DDiagram bottomRight,
+			float borderThickness) {
 		super(color, value, topLeft, bottomRight, borderThickness);
 	}
